+++ node/resolvers/directives/withPermissions.ts
@@ -4,6 +4,16 @@ import type { GraphQLField } from 'graphql'
 import { defaultFieldResolver } from 'graphql'
 import { SchemaDirectiveVisitor } from 'graphql-tools'
 
+import type StorefrontPermissions from '../../clients/storefrontPermissions'
+
+export const getUserPermission = async (
+  storefrontPermissions: StorefrontPermissions
+) => {
+  const result = await storefrontPermissions.checkUserPermission()
+
+  return result?.data?.checkUserPermission ?? null
+}
+
 export class WithPermissions extends SchemaDirectiveVisitor {
   public visitFieldDefinition(field: GraphQLField<any, any>) {
     const { resolve = defaultFieldResolver } = field
@@ -21,21 +31,18 @@ export class WithPermissions extends SchemaDirectiveVisitor {
 
       const appClients = context.vtex as any
 
-      appClients.storefrontPermissions = await storefrontPermissions
-        .checkUserPermission()
-        .then((result: any) => {
-          return result?.data?.checkUserPermission ?? null
-        })
-        .catch((error: any) => {
-          if (!adminUserAuthToken) {
-            logger.error({
-              message: 'getPermissionsError',
-              error,
-            })
-          }
+      appClients.storefrontPermissions = await getUserPermission(
+        storefrontPermissions
+      ).catch((error: any) => {
+        if (!adminUserAuthToken) {
+          logger.error({
+            message: 'getPermissionsError',
+            error,
+          })
+        }
 
-          return null
-        })
+        return null
+      })
 
       return resolve(root, args, context, info)
     }